codegen: Don't free value if property setter takes ownership
Correctly handle owned property accessor in object initializer.

In addition to c0e955db075d3d155782c167a0abb81e0dce5f59

See https://gitlab.gnome.org/GNOME/vala/issues/953

+++ codegen/valaccodebasemodule.vala
@@ -5145,7 +5145,7 @@ public abstract class Vala.CCodeBaseModule : CodeGenerator {
 					inst_ma.target_value = typed_inst;
 					store_property (p, inst_ma, init.initializer.target_value);
 					// FIXME Do not ref/copy in the first place
-					if (requires_destroy (init.initializer.target_value.value_type)) {
+					if (!p.set_accessor.value_type.value_owned && requires_destroy (init.initializer.target_value.value_type)) {
 						ccode.add_expression (destroy_value (init.initializer.target_value));
 					}
 				}
